+++ src/map/status_effect.h
@@ -783,7 +783,7 @@ public:
     void SetType(uint16 Type);
     void SetSlot(uint8 Slot);
     void SetIcon(uint16 Icon);
-    void SetPower(uint16 Power);
+    void SetPower(uint32 Power);
     void SetSubPower(uint16 subPower);
     void SetTier(uint16 tier);
     void SetDuration(uint32 Duration);
@@ -813,7 +813,7 @@ private:
     EFFECT m_StatusID{ EFFECT_NONE }; // основной тип эффекта
     uint32 m_SubID{ 0 };              // дополнительный тип эффекта
     uint16 m_Icon{ 0 };               // иконка эффекта
-    uint16 m_Power{ 0 };              // сила эффекта
+    uint32 m_Power{ 0 };              // сила эффекта
     uint16 m_SubPower{ 0 };           // Secondary power of the effect
     uint16 m_Tier{ 0 };               // Tier of the effect
     uint32 m_Flag{ 0 };               // флаг эффекта (условия его исчезновения)
